+++ packages/terra-application-utility/src/utility/_UtilityButton.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
-import { injectIntl, intlShape } from 'react-intl';
+import { injectIntl } from 'react-intl';
 import IconChevronDown from 'terra-icon/lib/icon/IconChevronDown';
 import IconChevronRight from 'terra-icon/lib/icon/IconChevronRight';
 import UtilityMenu from './_UtilityMenu';
@@ -27,9 +27,10 @@ const propTypes = {
    */
   onDisclose: PropTypes.func.isRequired,
   /**
+   * @private
    * The intl object to be injected for translations. Provided by the injectIntl function.
    */
-  intl: intlShape.isRequired,
+  intl: PropTypes.shape({ formatMessage: PropTypes.func }).isRequired,
   /**
    * The key of the top level menu page.
    */