+++ packages/terra-date-picker/src/react-datepicker/day.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames/bind'
-import { injectIntl, intlShape } from 'react-intl';
+import { injectIntl } from 'react-intl';
 import VisuallyHiddenText from 'terra-visually-hidden-text';
 import {
   getDay,
@@ -49,7 +49,7 @@ class Day extends React.Component {
      * @private
      * Internationalization object with translation APIs. Provided by `injectIntl`.
      */
-    intl: intlShape,
+    intl: PropTypes.shape({ formatMessage: PropTypes.func }),
     /**
      * Month value for the date entered.
      */